Preemptive start on traditional API server with LiteDB
Due to constant updates leading to the need to refresh cahce, I'm getting ready to move to a API server to handle fetching data. I'll be using LiteDB as a database.

For better reproducablity, this update also brings with it Git-based versions that include a partial has in the footer.

Aside from that, this update is mostly brings with it a lot of internal reorganization change. Most of the noticable changes before this point were added earlier because a lot of the cruical parts were already turned into modular components.