Fpcis on making into app
Went back to the original Blazor template to focus on making into a proper web app with PWA support. My original intentions was to turn this into a website based on web 1.0 ideologies. While it is possible, I wasn't doing myself any favors in making work with mobile because I'm still learning Bootstrap. It was becoming more trouble then it was worth that I decided to do a soft reboot.