Overhauled layout again
Overhauled layout for the third time now. This time I ported over the nav bar and footer from the ASP.NET Web App template instead of trying to rewrite the whole thing from scratch. The dark and light themes for the bottom and top rows have also been customized. The dark blue from the original Blazor template is now part of the dark theme, respectfully, while the light theme now has it's own colors. While everything else remains the same, there is a bug with the drop down menu when on mobile that I'm trying to figure out.