Preliminary support for comics
Comics are now supported using the same gallery and artwork API as the submissions. This was done by making the most important parts of the submission part of a base that's inherited by both the Artwork and Comics models. Meanwhile, the artwork property in the gallery model was turned into a generic that accepts only a class, making it backwards compatible with the existing API. Everything else follows the same principles introduced in the initial rollout of the gallery API.

- Advertisements was moved to the bottom